Jason's animation events. Also reworked entity-types system.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@2209 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/Model.h
@@ -102,7 +102,7 @@ public:
 
 	// load raw animation frame animation from given file, and build a 
 	// animation specific to this model
-	CSkeletonAnim* BuildAnimation(const char* filename,float speed,size_t actionpos);
+	CSkeletonAnim* BuildAnimation(const char* filename,float speed, double actionpos);
 
 	// add a prop to the model on the given point
 	void AddProp(SPropPoint* point,CModel* model);
